Extract buildArgumentClientContext to pure utils module
Moving the function to ats-feedback.utils.js (no imports) breaks the
Redis-at-import-time chain: ats-feedback.adapter.js pulls in
case-document-content-helper.js → db/redis.js, which opens a connection
at import time and prevents the node:test process from exiting cleanly.
The test now imports from the pure utils module; the adapter re-exports
the function for external callers.
