+++ ethicapp/backend/external-services/adapters/__tests__/ats-feedback.adapter.node-test.mjs
@@ -1,6 +1,6 @@
 import assert from "node:assert/strict";
 import test from "node:test";
-import { buildArgumentClientContext } from "../ats-feedback.adapter.js";
+import { buildArgumentClientContext } from "../ats-feedback.utils.js";
 
 // ─── buildArgumentClientContext ───────────────────────────────────────────────
